+++ CHANGELOG.md
@@ -6,9 +6,14 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.218 ] - 2025/06/24
+
+  * Updated `SpotifyConnectDirectoryTask.ActivateCastAppSpotify` method to ensure that the SpotifyAppTask is activated on the group leader if the device name is a cast group.
+
 ###### [ 1.0.217 ] - 2025/06/23
 
   * Updated underlying `PyChromecast` package requirement from version `>=14.0.5` to `>=14.0.7`.
+  * Updated `SpotifyConnectZeroconfCastController.GetInformation` method to back out the Cast `getInfo.deviceAPI_isGroup` parameter changes from v1.0.216.  This seems to have caused more issues than it prevents.
 
 ###### [ 1.0.216 ] - 2025/06/21